bugfix: generalize mongodb authentication
