+++ lib/storage/metadata/mongoclient/LogConsumer.js
@@ -2,6 +2,7 @@
 
 const MongoClient = require('mongodb').MongoClient;
 const ListRecordStream = require('./ListRecordStream');
+const MongoUtils = require('./utils');
 
 /**
  * @class
@@ -16,8 +17,9 @@ class LogConsumer {
      * @param {string} logger - logger
      */
     constructor(mongoConfig, logger) {
-        const { replicaSetHosts, database } = mongoConfig;
-        this._mongoUrl = `mongodb://${replicaSetHosts}/local`;
+        const { authCredentials, replicaSetHosts, database } = mongoConfig;
+        const cred = MongoUtils.credPrefix(authCredentials);
+        this._mongoUrl = `mongodb://${cred}${replicaSetHosts}/local`;
         this._logger = logger;
         this._oplogNsRegExp = new RegExp(`^${database}\\.`);
         // oplog collection
