+++ lib/storage/metadata/mongoclient/MongoClientInterface.js
@@ -89,14 +89,7 @@ class MongoClientInterface {
         const { replicaSetHosts, writeConcern, replicaSet, readPreference, path,
         database, logger, replicationGroupId, authCredentials, config }
         = params;
-        let cred = '';
-        if (authCredentials &&
-            authCredentials.username &&
-            authCredentials.password) {
-            const username = encodeURIComponent(authCredentials.username);
-            const password = encodeURIComponent(authCredentials.password);
-            cred = `${username}:${password}@`;
-        }
+        const cred = MongoUtils.credPrefix(authCredentials);
         this.mongoUrl = `mongodb://${cred}${replicaSetHosts}/` +
             `?w=${writeConcern}&replicaSet=${replicaSet}` +
             `&readPreference=${readPreference}`;
