+++ lib/storage/metadata/mongoclient/utils.js
@@ -27,4 +27,24 @@ function unserialize(objMD) {
     }
 }
 
-module.exports = { escape, unescape, serialize, unserialize };
+function credPrefix(authCredentials) {
+    let cred = '';
+
+    if (authCredentials &&
+        authCredentials.username &&
+        authCredentials.password) {
+        const username = encodeURIComponent(authCredentials.username);
+        const password = encodeURIComponent(authCredentials.password);
+        cred = `${username}:${password}@`;
+    }
+
+    return cred;
+}
+
+module.exports = {
+    credPrefix,
+    escape,
+    serialize,
+    unescape,
+    unserialize,
+};
